+++ src/js/components/Exporter.js
@@ -1,12 +1,13 @@
 import xs from "xstream"
 import { div, i, ul, li, p, input, button, span, a } from "@cycle/dom"
-import { isEmpty, mergeLeft, equals } from "ramda"
+import { isEmpty, mergeLeft, equals, keys, toUpper, all, identity } from "ramda"
 import { loggerFactory } from "../utils/logger"
 import delay from "xstream/extra/delay"
 import debounce from "xstream/extra/debounce"
 import sampleCombine from "xstream/extra/sampleCombine"
 import dropRepeats from "xstream/extra/dropRepeats"
-import { convertToCSV } from "../utils/export"
+import { convertToCSV, convertTableToMd, convertFilterToMd, convertSelectedSamplesToMd } from "../utils/export"
+import { map } from "jquery"
 
 /**
  * @module components/Exporter
@@ -27,6 +28,7 @@ function intent(domSource$) {
   const exportPlotsTrigger$ = domSource$.select(".export-clipboard-plots").events("click")
   const exportHeadTableTrigger$ = domSource$.select(".export-clipboard-headTable").events("click")
   const exportTailTableTrigger$ = domSource$.select(".export-clipboard-tailTable").events("click")
+  const exportMdReportTrigger$ = domSource$.select(".export-clipboard-mdReport").events("click")
 
   const modalTrigger$ = domSource$.select(".modal-open-btn").events("click")
   const modalCloseTrigger$ = domSource$.select(".export-close").events("click")
@@ -41,6 +43,7 @@ function intent(domSource$) {
     exportPlotsTrigger$: exportPlotsTrigger$,
     exportHeadTableTrigger$: exportHeadTableTrigger$,
     exportTailTableTrigger$: exportTailTableTrigger$,
+    exportMdReportTrigger$: exportMdReportTrigger$,
     modalTrigger$: modalTrigger$,
     modalCloseTrigger$: modalCloseTrigger$,
     testTrigger$: testTrigger$,
@@ -80,6 +83,18 @@ function model(actions, state$, vega$, config) {
   const headTablePresent$ = state$.map((state) => notEmptyOrUndefined(state.headTable?.data)).startWith(false)
   const tailTablePresent$ = state$.map((state) => notEmptyOrUndefined(state.tailTable?.data)).startWith(false)
 
+  const genericMdDataPresent$ = xs.combine(signaturePresent$, plotsPresent$, headTablePresent$, tailTablePresent$).map((arr) => all(identity)(arr))
+  const diseaseMdDataPresent$ = xs.combine(plotsPresent$, headTablePresent$, tailTablePresent$).map((arr) => all(identity)(arr))
+  const correlationMdDataPresent$ = plotsPresent$
+
+  const mdPresentSelector = {
+    "": genericMdDataPresent$, // generic treatment
+    "DISEASE": diseaseMdDataPresent$,
+    "CORRELATION": correlationMdDataPresent$,
+  }
+
+  const mdReportPresent$ = (mdPresentSelector[toUpper(config.workflowName)] ?? mdPresentSelector[""])
+
   const url$ = state$.map((state) => state.routerInformation.pageStateURL).startWith("")
   const signature$ = state$.map((state) => state.form.signature?.output).startWith("")
   // result already contains 'data:image/png;base64,'
@@ -99,10 +114,104 @@ function model(actions, state$, vega$, config) {
     .map((data) => convertToCSV(data))
     .startWith("")
 
+  const urlMd$ = url$.map((url) => "["+url+"]("+url+")").startWith("")
+
+  const treatmentQueryMd$ = state$.map((state) => state.form.check?.output) // generic treatment WF
+  const signatureQueryMd$ = state$.map((state) => state.form.query) // disease WF
+  const signatureQuery1Md$ = state$.map((state) => state.form.query1) // correlation WF
+  const signatureQuery2Md$ = state$.map((state) => state.form.query2) // correlation WF
+
+  // present in generic treatment WFs
+  const samplesMd$ = state$.map((state) => convertSelectedSamplesToMd(state.form.sampleSelection?.data)).startWith("")
+  const signatureMd$ = signature$
+
+  const filterMd$ = state$.map((state) => convertFilterToMd(state.filter.filter_output, state.settings.filter.values))
+    .startWith("")
+
+  const plotMd$ = plotFile$
+    .filter((data) => data != "")
+    .map((data) => "![](" + data + ")")
+    .startWith("")
+
+  const headTableMd$ = state$.map((state) => state.headTable?.data)
+    .filter((data) => notEmptyOrUndefined(data))
+    .map((data) => convertTableToMd(data))
+    .startWith("")
+  
+  const tailTableMd$ = state$.map((state) => state.tailTable?.data)
+    .filter((data) => notEmptyOrUndefined(data))
+    .map((data) => convertTableToMd(data))
+    .startWith("")
+
+  // data streams with headers added
+  const addHeaderL2 = (h) => (s) => ("\n" + "## " + h + "\n\n" + s)
+  const urlMdWH$ = urlMd$.map(addHeaderL2("Search query URL"))
+  const treatmentQueryMdWH$ = treatmentQueryMd$.map(addHeaderL2("Treatment query"))
+  const signatureQueryMdWH$ = signatureQueryMd$.map(addHeaderL2("Signature query"))
+  const signatureQuery1MdWH$ = signatureQuery1Md$.map(addHeaderL2("Signature query 1"))
+  const signatureQuery2MdWH$ = signatureQuery2Md$.map(addHeaderL2("Signature query 2"))
+  const samplesMdWH$ = samplesMd$.map(addHeaderL2("Selected samples"))
+  const signatureMdWH$ = signatureMd$.map(addHeaderL2("Signature"))
+  const filterMdWH$ = filterMd$.map(addHeaderL2("Filter"))
+  const plotMdWH$ = plotMd$.map(addHeaderL2("Plot"))
+  const headTableMdWH$ = headTableMd$.map(addHeaderL2("Top table"))
+  const tailTableMdWH$ = tailTableMd$.map(addHeaderL2("Bottom table"))
+
+  const WFTitleMd$ = xs.of("# " + config.workflowName + " Workflow report")
+
+  const genericMd$ = xs.combine(WFTitleMd$, urlMdWH$, treatmentQueryMdWH$, samplesMdWH$, signatureMdWH$, filterMdWH$, plotMdWH$, headTableMdWH$, tailTableMdWH$)
+    .map(([WFTitle, url, treatmentQuery, samples, signature, filter, plot, head, tail]) => 
+      [
+        WFTitle,
+        url,
+        treatmentQuery,
+        samples,
+        signature,
+        filter,
+        plot,
+        head,
+        tail
+      ].join("\n")
+    )
+
+  const DiseaseMd$ = xs.combine(WFTitleMd$, urlMdWH$, signatureQueryMdWH$, filterMdWH$, plotMdWH$, headTableMdWH$, tailTableMdWH$)
+    .map(([WFTitle, url, signatureQuery, filter, plot, head, tail]) => 
+      [
+        WFTitle,
+        url,
+        signatureQuery,
+        filter,
+        plot,
+        head,
+        tail
+      ].join("\n")
+    )
+
+  const CorrelationMd$ = xs.combine(WFTitleMd$, urlMdWH$, signatureQuery1MdWH$, signatureQuery2MdWH$, filterMdWH$, plotMdWH$)
+    .map(([WFTitle, url, signatureQuery1, signatureQuery2, filter, plot, head, tail]) => 
+      [
+        WFTitle,
+        url,
+        signatureQuery1,
+        signatureQuery2,
+        filter,
+        plot,
+      ].join("\n")
+    )
+
+  const MdSelector = {
+    "": genericMd$, // generic treatment
+    "DISEASE": DiseaseMd$,
+    "CORRELATION": CorrelationMd$,
+  }
+
+  const selectedMd$ = MdSelector[toUpper(config.workflowName)] ?? MdSelector[""]
+
   const urlFile$ = url$.map(url => "data:text/plain;charset=utf-8," + url)
   const signatureFile$ = signature$.map(signature => "data:text/plain;charset=utf-8," + signature)
   const headTableCsvFile$ = headTableCsv$.map(headTableCsv => "data:text/tsv;charset=utf-8," + encodeURIComponent(headTableCsv))
   const tailTableCsvFile$ = tailTableCsv$.map(tailTableCsv => "data:text/tsv;charset=utf-8," + encodeURIComponent(tailTableCsv))
+  const mdReportFile$ = selectedMd$.map(md => "data:text/plain;charset=utf-8," + encodeURIComponent(md))
 
   const clipboardLinkFab$ = actions.exportLinkTriggerFab$
     .compose(sampleCombine(url$))
@@ -172,21 +281,21 @@ function model(actions, state$, vega$, config) {
     }))
     .remember()
 
+  const clipboardMdReport$ = actions.exportMdReportTrigger$
+    .compose(sampleCombine(selectedMd$))
+    .map(([_, md]) => ({
+      sender: "mdReport",
+      data: md,
+    }))
+    .remember()
+
   const testAction$ = actions.testTrigger$
-    .compose(sampleCombine(plotFile$))
-    .map(([_, data]) => {
-      // input data is "data:image/png;base64,abcdef0123456789..."
-      const parts = data.split(';base64,');
-      const imageType = parts[0].split(':')[1];
-      const decodedData = window.atob(parts[1]);
-      const uInt8Array = new Uint8Array(decodedData.length);
-      for (let i = 0; i < decodedData.length; i++) {
-        uInt8Array[i] = decodedData.charCodeAt(i);
-      }
-      const blob = new Blob([uInt8Array], { type: imageType })
+    .compose(sampleCombine(selectedMd$))
+    .map(([_, md]) => {
+     
       return {
-        type: imageType,
-        data: blob,
+        sender: 'test-fab',
+        data: md,
       }
     })
     .remember()
@@ -194,12 +303,23 @@ function model(actions, state$, vega$, config) {
   return {
     reducers$: xs.empty(),
     modal$: xs.merge(openModal$, closeModal$),
-    clipboard$: xs.merge(clipboardLinkFab$, clipboardSignatureFab$, clipboardLink$, clipboardSignature$, clipboardPlots$, clipboardHeadTable$, clipboardTailTable$, testAction$),
+    clipboard$: xs.merge(
+      clipboardLinkFab$,
+      clipboardSignatureFab$,
+      clipboardLink$,
+      clipboardSignature$,
+      clipboardPlots$,
+      clipboardHeadTable$,
+      clipboardTailTable$,
+      clipboardMdReport$,
+      testAction$,
+    ),
     dataPresent: {
       signaturePresent$: signaturePresent$,
       plotsPresent$: plotsPresent$,
       headTablePresent$: headTablePresent$,
       tailTablePresent$: tailTablePresent$,
+      mdReportPresent$: mdReportPresent$,
     },
     exportData: {
       url$: url$,
@@ -208,6 +328,7 @@ function model(actions, state$, vega$, config) {
       plotFile$: plotFile$,
       headTableCsvFile$: headTableCsvFile$,
       tailTableCsvFile$: tailTableCsvFile$,
+      mdReportFile$: mdReportFile$,
     }
   }
 }
@@ -268,12 +389,14 @@ function view(state$, dataPresent, exportData, config, clipboard) {
         dataPresent.plotsPresent$,
         dataPresent.headTablePresent$,
         dataPresent.tailTablePresent$,
+        dataPresent.mdReportPresent$,
         exportData.url$,
         exportData.urlFile$,
         exportData.signatureFile$,
         exportData.plotFile$,
         exportData.headTableCsvFile$,
         exportData.tailTableCsvFile$,
+        exportData.mdReportFile$,
         clipboard.copyImagesPermission$,
         clipboardResultAutoClear$,
       )
@@ -282,12 +405,14 @@ function view(state$, dataPresent, exportData, config, clipboard) {
         plotsPresent,
         headTablePresent,
         tailTablePresent,
+        mdReportPresent,
         url,
         urlFile,
         signatureFile,
         plotFile,
         headTableCsvFile,
         tailTableCsvFile,
+        mdReportFile,
         clipboardPermissions,
         clipboardResult,
       ]) => {
@@ -337,6 +462,7 @@ function view(state$, dataPresent, exportData, config, clipboard) {
               addExportDiv("plot", "Copy " + config.plotName + " plot", ".export-clipboard-plots", plotFile, "plot.png", plotsPresent, copyImagesPermission),
               addExportDiv("headTable", "Copy top table", ".export-clipboard-headTable", headTableCsvFile, "table.tsv", headTablePresent),
               addExportDiv("tailTable", "Copy bottom table", ".export-clipboard-tailTable", tailTableCsvFile, "table.tsv", tailTablePresent),
+              addExportDiv("mdReport", "Copy MarkDown report", ".export-clipboard-mdReport", mdReportFile, "report.md", mdReportPresent),
               // div(".row", [
               //   span(".col .s6 .push-s1", "Export report"),
               //   span(".btn .col .s1 .offset-s3 .export-file-report .disabled", i(".material-icons", "file_download")),
@@ -375,6 +501,7 @@ function Exporter(sources) {
     plotName: "binned similarity", // part of the text to be displayed for plot copy/download
     fabSignature: "update", // part of FAB class name, set to "", ".hide" or ".disabled". 
                             //"update" sets ".disabled" when the signature is not available and updates the FAB when it becomes available
+    workflowName: "", // Both the name to add in the MarkDown report and select how the content in combined
   }
   const fullConfig = mergeLeft(sources.config, defaultConfig)